requirements.txt
python-dotenv
python-telegram-bot
google-generativeai
httpx
flask
pillow